requirements.txt
# F&O Trading System Dependencies

# HTTP client for API calls
aiohttp>=3.9.0
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Logging
colorlog>=6.8.0

# Date/time handling
python-dateutil>=2.8.2

# Optional: For better JSON handling
orjson>=3.9.0

# Web dashboard
flask>=3.0.0
